Move value cacher to a new file. Improve logic for refreshing the records.
